server/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
elasticsearch==8.10.1
python-telegram-bot==20.6
pyyaml==6.0.1
python-dotenv==1.0.0
pydantic==2.4.2 